feat(lco): per-station Csound-GEN author replaces the lexicon recipe author
The LCO recipe author is now the coder LLM validated in the PoC: it plans 2-3
timbre STATIONS and emits one Csound GEN spectrum per station;
backend/lco_author.py maps each spectrum to an Additive keyframe and the station
order to a Motion trajectory that the shipping DcoBaker bakes. Replaces the
dco_recipe lexicon/S2 author at the same "dco" IPC mode -- C++ transport +
recipe parse unchanged. dco_recipe.py stays in-tree, now unused.

GEN -> partials is DIRECT and exact (test Part A: 0.00e+00 over 200 trials):
GEN10 arg k -> (h=k, a=strength, phase=0); GEN09 triple -> (h=partial#,
a=strength, phase=radians). GEN09's non-integer partial# is INHARMONIC, which
needs the one guardian-adjacent change:

  DcoBaker::renderAdditive now takes a FLOAT harmonic h (Partial::h int->float,
  clamped [0, kMaxHarmonics], NaN-guarded). No behavior change for existing
  data -- every existing recipe/old-author partial has integer h>=1 (dco_recipe
  floored h to [1,1024]; DCO presets aren't persisted yet). Held-note
  crossfade-follow invariant untouched: only renderAdditive's inner loop
  changed; bake()/morph/normalize/WavetableOscillator paths are unchanged.
  Verified by adversarial review + a DcoBaker-faithful bake harness.

Full author -> faithful bake reproduces the approved PoC: 8/8 distinct, 0
static, centroid-sweep byte-identical to the validated run.

Coder coexists with the translator (run_instruct's (dir,device) cache);
_resolve_coder_model_dir auto-discovers <root>/coder|lco-coder/*. Response keeps
the C++-read fields (resolved.adjectives = stations; flags empty) so PromptPanel
is unchanged -- the station-as-reading disclosure is Phase 3.

Harness: tools/test_lco_author.py (Part A exactness, Part B end-to-end).

+++ src/dsp/DcoRecipe.h
@@ -17,8 +17,13 @@
  */
 namespace dco {
 
-// One harmonic partial: harmonic number (1-based), amplitude, phase (radians).
-struct Partial { int h = 1; float a = 0.0f; float phase = 0.0f; };
+// One partial: harmonic number, amplitude, phase (radians). h is a FLOAT: an
+// integer h is an ordinary harmonic; a non-integer h is an INHARMONIC partial
+// (glassy/metallic/bell — a Csound GEN09 spectrum), which DcoBaker renders as a
+// sub-cycle sine that the engine's per-pitch band-limited mips resolve. An
+// integer h >= 1 (all existing recipe data) renders exactly as the former int
+// field did; non-integer h is the new inharmonic case.
+struct Partial { float h = 1.0f; float a = 0.0f; float phase = 0.0f; };
 
 // A keyframe = one single-cycle spectrum with a closed-form param->spectrum map.
 // The field ranges noted below are DcoBaker's contract — it clamps every field
